pyproject.toml
[tool.poetry]
name = "ramos-uc"
version = "0.1.0"
description = "Plataforma para la visualización y búsqueda de ramos de la Pontificia Universidad Católica."
authors = ["Nicolás Mc Intyre <nicomcintyrew@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.23"
django-allauth = "^0.63.6"
django-dotenv = "^1.4.2"
django-manifest-loader = "^1.0.0"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
django-background-tasks = "^1.2.5"
requests = "^2.32.3"
pyjwt = "2.10"
cryptography = "^44.0.0"

[tool.poetry.dev-dependencies]
black = "^22.12"
flake8 = "^3.9.2"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
